examples/layout-tests/serviceworker/current-on-reload.html

<!DOCTYPE html>
<title>Service Worker: Current on reload</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/testutils.js"></script>
<body>
<script>
var t = async_test('current is set upon reload after registration');
t.step(function() {
    var reloaded = false;
    var scope = 'resources/blank.html';
    navigator.serviceWorker.unregister(scope).then(doTest, doTest);

    function doTest() {
        withIframe(scope, t.step_func(onIframeLoad));
    }

    function onIframeLoad(frame) {
        var w = frame.contentWindow;

        function onActive(event) {
           assert_equals(w.navigator.serviceWorker.current, null,
                         'current should be null until the document is reloaded');
           reloaded = true;
           w.location.reload();
        }

        function onRegister(worker) {
            // FIXME: Remove this check after crbug.com/365252 is fixed.
            if (worker.state == 'active')
                t.step(onActive);
            worker.addEventListener('statechange', function(event) {
                if (event.target.state == 'active')
                    t.step(onActive);
            });
        }

        if (reloaded) {
            assert_true(w.navigator.serviceWorker.current instanceof ServiceWorker,
                        'current should be a ServiceWorker object upon reload');
            t.done();
            return;
        }

        w.navigator.serviceWorker.register(
            'worker-no-op.js', {scope: 'blank.html'}
        ).then(t.step_func(onRegister), t.step_func(function(reason) {
            assert_unreached('Registration should succeed, but failed: ' + reason.name);
        }));
    }
});
</script>
</body>
